raspberrypi/pyproject.toml
[project]
name = "raspberrypi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "colorama>=0.4.6",
    "httpx>=0.28.1",
    "mcp[cli]>=1.6.0",
    "paramiko>=3.5.1",
    "python-dotenv>=1.1.0",
]
